refactor(desktop): 启动装配模块归入 main/bootstrap/
把 index 的各启动/装配模块归到 main/bootstrap/ 子目录并加桶导出，index 作为组合根从 ./bootstrap 取用：
- 移入：connections-runtime / window / poller / repo-mirror / process-tweaks / update-runner /
  pragent-runtime（相对 import 同步 ../ 提一级到 adapters / services / utils）。
- 新增 bootstrap/index.ts 桶导出。index 从 './bootstrap/index.js' 统一引入。

纯移动 / 分组、行为不变；desktop typecheck/lint/build 全绿。
