+++ apps/desktop/src/main/bootstrap/process-tweaks.ts
